Reduced current consumption by using LPM3.5
Sleep current measured as 420nA in LPM3.5, which is down from 1.5uA in LPM3 used previously. Many variables have been moved to FRAM, using the PERSISTENT pragma and the intiailisation of these have had to change. BKMEM is used to store a variable that indicates whether or not initialisation can be skipped or not.