Added turbo mode
In turbo mode, temperature and humidity are sampled at 1Hz. Also added a 3 second delay at startup, before measuring the battery voltage. Finally nt3h_check_address has been modified to return 1 when the capability container is not as expected. The capability container must not be overwritten before checking for a configuration text record. Overwriting the CC in page0 seems to change the NDEF text record in page 1.